+++ test/fixtures/script.js
@@ -4,5 +4,5 @@ var oldlib = require('./old-lib')
 run()
 
 function run() {
-  oldlib.old()
+  oldlib.oldfunction()
 }